+++ code/render/coregraphics/vk/vkmemoryindexbufferloader.cc
@@ -40,7 +40,7 @@ VkMemoryIndexBufferLoader::OnLoadRequested()
 		VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO,
 		NULL,
 		0,					// use for sparse buffers
-		this->numIndices * IndexType::SizeOf(this->indexType),
+		(uint32_t)(this->numIndices * IndexType::SizeOf(this->indexType)),
 		VK_BUFFER_USAGE_INDEX_BUFFER_BIT,
 		VK_SHARING_MODE_EXCLUSIVE,						// can only be accessed from the creator queue,
 		1,												// number of queues in family
