+++ code/render/coregraphics/vk/vkmemoryvertexbufferloader.cc
@@ -40,7 +40,7 @@ VkMemoryVertexBufferLoader::OnLoadRequested()
 		VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO,
 		NULL,
 		0,					// use for sparse buffers
-		vertexSize * this->numVertices,
+		(uint32_t)(vertexSize * this->numVertices),
 		VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
 		VK_SHARING_MODE_EXCLUSIVE,						// can only be accessed from the creator queue,
 		1,												// number of queues in family
